+++ integration/test_batch_v4.py
@@ -98,6 +98,7 @@ def client_factory(
             ],
             references=[ReferenceProperty(name="test", target_collection=name_fixture)],
             multi_tenancy_config=Configure.multi_tenancy(multi_tenant),
+            vectorizer_config=Configure.Vectorizer.none(),
         )
         return client_fixture, name_fixture
 
@@ -536,3 +537,12 @@ def test_error_reset(client_factory: ClientFactory) -> None:
     assert len(errs) == 1
     assert errs[0].object_.properties is not None
     assert errs[0].object_.properties["name"] == 1
+
+
+def test_non_existant_collection(client_factory: ClientFactory) -> None:
+    client, _ = client_factory()
+    with client.batch.dynamic() as batch:
+        batch.add_object(properties={"name": 2}, collection="DoesNotExist")
+
+    # above should not throw - depending on the autoschema config this might create an error or
+    # not, so we do not check for errors here