+++ integration/test_collection_openai.py
@@ -525,7 +525,7 @@ def test_near_vector_generate_and_group_by_with_everything(
     assert list(res.groups.values())[1].generated == "No"
 
 
-def test_openapi_invalid_key(request: SubRequest) -> None:
+def test_openai_invalid_key(request: SubRequest) -> None:
     local_client = weaviate.connect_to_local(
         port=8086, grpc_port=50057, headers={"X-OpenAI-Api-Key": "IamNotValid"}
     )
@@ -535,13 +535,14 @@ def test_openapi_invalid_key(request: SubRequest) -> None:
         name=request.node.name,
         properties=[Property(name="text", data_type=DataType.TEXT)],
         generative_config=Configure.Generative.openai(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     collection.data.insert(properties={"text": "test"})
     with pytest.raises(WeaviateQueryError):
         collection.generate.fetch_objects(single_prompt="tell a joke based on {text}")
 
 
-def test_openapi_no_module(request: SubRequest) -> None:
+def test_openai_no_module(request: SubRequest) -> None:
     local_client = weaviate.connect_to_local(
         port=8080, grpc_port=50051, headers={"X-OpenAI-Api-Key": "doesnt matter"}
     )
@@ -551,6 +552,7 @@ def test_openapi_no_module(request: SubRequest) -> None:
         name=request.node.name,
         properties=[Property(name="text", data_type=DataType.TEXT)],
         generative_config=Configure.Generative.openai(),
+        vectorizer_config=Configure.Vectorizer.none(),
     )
     collection.data.insert(properties={"text": "test"})
     with pytest.raises(WeaviateQueryError):