+++ weaviate/collections/classes/config_named_vectors.py
@@ -28,6 +28,7 @@ from weaviate.collections.classes.config_vectorizers import (
     _Text2VecHuggingFaceConfigCreate,
     _Text2VecJinaConfigCreate,
     _Text2VecOctoConfig,
+    _Text2VecOllamaConfig,
     _Text2VecOpenAIConfigCreate,
     _Text2VecPalmConfigCreate,
     _Text2VecTransformersConfigCreate,
@@ -196,10 +197,10 @@ class _NamedVectors:
         source_properties: Optional[List[str]] = None,
         vector_index_config: Optional[_VectorIndexConfigCreate] = None,
         vectorize_collection_name: bool = True,
-        model: Optional[Union[OpenAIModel, str]] = None,
+        model: Optional[str] = None,
         base_url: Optional[str] = None,
     ) -> _NamedVectorConfigCreate:
-        """Create a named vector using the `text2vec_octoai` model.
+        """Create a named vector using the `text2vec-octoai` model.
 
         See the [documentation](https://weaviate.io/developers/weaviate/modules/retriever-vectorizer-modules/text2vec-octoai)
         for detailed usage.
@@ -226,6 +227,49 @@ class _NamedVectors:
             source_properties=source_properties,
             vectorizer=_Text2VecOctoConfig(
                 baseURL=base_url,
+                model=model,
+                vectorizeClassName=vectorize_collection_name,
+            ),
+            vector_index_config=vector_index_config,
+        )
+
+    @staticmethod
+    def text2vec_ollama(
+        name: str,
+        *,
+        source_properties: Optional[List[str]] = None,
+        vector_index_config: Optional[_VectorIndexConfigCreate] = None,
+        vectorize_collection_name: bool = True,
+        model: Optional[str] = None,
+        api_endpoint: Optional[str] = None,
+    ) -> _NamedVectorConfigCreate:
+        """Create a named vector using the `text2vec-ollama` model.
+
+        See the [documentation](https://weaviate.io/developers/weaviate/modules/retriever-vectorizer-modules/text2vec-ollama)
+        for detailed usage.
+
+        Arguments:
+            `name`
+                The name of the named vector.
+            `source_properties`
+                Which properties should be included when vectorizing. By default all text properties are included.
+            `vector_index_config`
+                The configuration for Weaviate's vector index. Use wvc.config.Configure.VectorIndex to create a vector index configuration. None by default
+            `vectorize_collection_name`
+                Whether to vectorize the collection name. Defaults to `True`.
+            `model`
+                The model to use. Defaults to `None`, which uses the server-defined default.
+            `vectorize_collection_name`
+                Whether to vectorize the collection name. Defaults to `True`.
+            `api_endpoint`
+                The base URL to use where API requests should go. Defaults to `None`, which uses the server-defined default.
+
+        """
+        return _NamedVectorConfigCreate(
+            name=name,
+            source_properties=source_properties,
+            vectorizer=_Text2VecOllamaConfig(
+                apiEndpoint=api_endpoint,
                 model=model,
                 vectorizeClassName=vectorize_collection_name,
             ),
@@ -682,6 +726,7 @@ class _NamedVectors:
         vectorize_collection_name: bool = True,
         api_endpoint: Optional[str] = None,
         model_id: Optional[str] = None,
+        title_property: Optional[str] = None,
     ) -> _NamedVectorConfigCreate:
         """Create a named vector using the `text2vec_palm` model.
 
@@ -705,6 +750,8 @@ class _NamedVectors:
                 The API endpoint to use without a leading scheme such as `http://`. Defaults to `None`, which uses the server-defined default
             `model_id`
                 The model ID to use. Defaults to `None`, which uses the server-defined default.
+            `title_property`
+                The Weaviate property name for the `gecko-002` or `gecko-003` model to use as the title.
 
         Raises:
             `pydantic.ValidationError` if `api_endpoint` is not a valid URL.
@@ -717,6 +764,7 @@ class _NamedVectors:
                 apiEndpoint=api_endpoint,
                 modelId=model_id,
                 vectorizeClassName=vectorize_collection_name,
+                titleProperty=title_property,
             ),
             vector_index_config=vector_index_config,
         )