+++ weaviate/connect/integrations.py
@@ -1,4 +1,5 @@
 from typing import Dict, Optional, cast
+
 from pydantic import BaseModel, ConfigDict, Field
 
 
@@ -16,7 +17,7 @@ class _IntegrationConfig(BaseModel):
 
 class _IntegrationConfigCohere(_IntegrationConfig):
     api_key: str = Field(serialization_alias="X-Cohere-Api-Key")
-    request_per_minute_embeddings: Optional[int] = Field(
+    requests_per_minute_embeddings: Optional[int] = Field(
         serialization_alias="X-Cohere-Ratelimit-RequestPM-Embedding"
     )
     base_url: Optional[str] = Field(serialization_alias="X-Cohere-Baseurl")
@@ -25,7 +26,7 @@ class _IntegrationConfigCohere(_IntegrationConfig):
 class _IntegrationConfigOpenAi(_IntegrationConfig):
     api_key: str = Field(serialization_alias="X-Openai-Api-Key")
     organization: Optional[str] = Field(serialization_alias="X-Openai-Organization")
-    request_per_minute_embeddings: Optional[int] = Field(
+    requests_per_minute_embeddings: Optional[int] = Field(
         serialization_alias="X-Openai-Ratelimit-RequestPM-Embedding"
     )
     tokens_per_minute_embeddings: Optional[int] = Field(
@@ -37,7 +38,7 @@ class _IntegrationConfigOpenAi(_IntegrationConfig):
 class _IntegrationConfigAWS(_IntegrationConfig):
     access_key: str = Field(serialization_alias="X-Aws-Access-Key")
     secret_key: str = Field(serialization_alias="X-Aws-Secret-Key")
-    request_per_minute_embeddings: Optional[int] = Field(
+    requests_per_minute_embeddings: Optional[int] = Field(
         serialization_alias="X-Aws-Ratelimit-RequestPM-Embedding"
     )
     tokens_per_minute_embeddings: Optional[int] = Field(
@@ -47,7 +48,7 @@ class _IntegrationConfigAWS(_IntegrationConfig):
 
 class _IntegrationConfigVoyage(_IntegrationConfig):
     api_key: str = Field(serialization_alias="X-Voyageai-Api-Key")
-    request_per_minute_embeddings: Optional[int] = Field(
+    requests_per_minute_embeddings: Optional[int] = Field(
         serialization_alias="X-Voyageai-Ratelimit-RequestPM-Embedding"
     )
     tokens_per_minute_embeddings: Optional[int] = Field(
@@ -58,18 +59,18 @@ class _IntegrationConfigVoyage(_IntegrationConfig):
 
 class _IntegrationConfigJina(_IntegrationConfig):
     api_key: str = Field(serialization_alias="X-Jinaai-Api-Key")
-    request_per_minute_embeddings: Optional[int] = Field(
+    requests_per_minute_embeddings: Optional[int] = Field(
         serialization_alias="X-Jinaai-Ratelimit-RequestPM-Embedding"
     )
     base_url: Optional[str] = Field(serialization_alias="X-Jinaai-Baseurl")
 
 
 class _IntegrationConfigOcto(_IntegrationConfig):
-    api_key: str = Field(serialization_alias="X-Octoai-Api-Key")
-    request_per_minute_embeddings: Optional[int] = Field(
-        serialization_alias="X-Octoai-Ratelimit-RequestPM-Embedding"
+    api_key: str = Field(serialization_alias="X-OctoAI-Api-Key")
+    requests_per_minute_embeddings: Optional[int] = Field(
+        serialization_alias="X-OctoAI-Ratelimit-RequestPM-Embedding"
     )
-    base_url: Optional[str] = Field(serialization_alias="X-Octoai-Baseurl")
+    base_url: Optional[str] = Field(serialization_alias="X-OctoAI-Baseurl")
 
 
 class Integrations:
@@ -78,11 +79,11 @@ class Integrations:
         *,
         api_key: str,
         base_url: Optional[str] = None,
-        request_per_minute_embeddings: Optional[int] = None
+        requests_per_minute_embeddings: Optional[int] = None
     ) -> _IntegrationConfig:
         return _IntegrationConfigCohere(
             api_key=api_key,
-            request_per_minute_embeddings=request_per_minute_embeddings,
+            requests_per_minute_embeddings=requests_per_minute_embeddings,
             base_url=base_url,
         )
 
@@ -90,14 +91,14 @@ class Integrations:
     def openai(
         *,
         api_key: str,
-        request_per_minute_embeddings: Optional[int] = None,
+        requests_per_minute_embeddings: Optional[int] = None,
         tokens_per_minute_embeddings: Optional[int] = None,
         organization: Optional[str] = None,
         base_url: Optional[str] = None
     ) -> _IntegrationConfig:
         return _IntegrationConfigOpenAi(
             api_key=api_key,
-            request_per_minute_embeddings=request_per_minute_embeddings,
+            requests_per_minute_embeddings=requests_per_minute_embeddings,
             tokens_per_minute_embeddings=tokens_per_minute_embeddings,
             organization=organization,
             base_url=base_url,
@@ -109,13 +110,13 @@ class Integrations:
     #     *,
     #     access_key: str,
     #     secret_key: str,
-    #     request_per_minute_embeddings: Optional[int] = None,
+    #     requests_per_minute_embeddings: Optional[int] = None,
     #     tokens_per_minute_embeddings: Optional[int] = None,
     # ) -> _IntegrationConfig:
     #     return _IntegrationConfigAWS(
     #         access_key=access_key,
     #         secret_key=secret_key,
-    #         request_per_minute_embeddings=request_per_minute_embeddings,
+    #         requests_per_minute_embeddings=requests_per_minute_embeddings,
     #         tokens_per_minute_embeddings=tokens_per_minute_embeddings,
     #     )
 
@@ -123,13 +124,13 @@ class Integrations:
     def voyageai(
         *,
         api_key: str,
-        request_per_minute_embeddings: Optional[int] = None,
+        requests_per_minute_embeddings: Optional[int] = None,
         tokens_per_minute_embeddings: Optional[int] = None,
         base_url: Optional[str] = None
     ) -> _IntegrationConfig:
         return _IntegrationConfigVoyage(
             api_key=api_key,
-            request_per_minute_embeddings=request_per_minute_embeddings,
+            requests_per_minute_embeddings=requests_per_minute_embeddings,
             tokens_per_minute_embeddings=tokens_per_minute_embeddings,
             base_url=base_url,
         )
@@ -138,12 +139,12 @@ class Integrations:
     def jinaai(
         *,
         api_key: str,
-        request_per_minute_embeddings: Optional[int] = None,
+        requests_per_minute_embeddings: Optional[int] = None,
         base_url: Optional[str] = None
     ) -> _IntegrationConfig:
         return _IntegrationConfigJina(
             api_key=api_key,
-            request_per_minute_embeddings=request_per_minute_embeddings,
+            requests_per_minute_embeddings=requests_per_minute_embeddings,
             base_url=base_url,
         )
 
@@ -151,11 +152,11 @@ class Integrations:
     def octoai(
         *,
         api_key: str,
-        request_per_minute_embeddings: Optional[int] = None,
+        requests_per_minute_embeddings: Optional[int] = None,
         base_url: Optional[str] = None
     ) -> _IntegrationConfig:
         return _IntegrationConfigOcto(
             api_key=api_key,
-            request_per_minute_embeddings=request_per_minute_embeddings,
+            requests_per_minute_embeddings=requests_per_minute_embeddings,
             base_url=base_url,
         )